Add test
Iterate on the testcase under test/.
The test will check whether the compilation successfully finishes
and whether the genrated assembly code is valid MIPS assembly.

However, it will not test the correctness of the assembly code.
The test cannot utilize given inputs and compare outputs to expected
ones. These functions may be implemented in the future.